Speech Event Detection/Archive/requirements.txt
tensorflow
keras
librosa
python_speech_features
numpy
scikit-learn
pandas
tqdm
matplotlib
joblib
praat-textgrids